week8/requirements.txt
apscheduler
lxml
newspaper
numpy
pandas
pika
pymongo
python-jsonrpc
python-dateutil
redis
requests
scipy
sklearn
tensorflow
watchdog
